Add support for creating and resizing the Write account
Since the Write account needs to be a PDA, add support for creating
and resizing it as necessary.  We’re potentially missing Truncate and
Free instructions but we can live without them for the time being.

Furthermore, remove the Copy instruction.  We don’t use it currently
and having support for it while details of creation and resizing of
the account aren’t yet established create unnecessary maintenance
burden.

Co-authored-by: Dhruv D Jain <dhruv@iamsizzling.com>