Improve WriteIter implementation
Firstly, add `MAX_CHUNK_SIZE` constant and use it as the default chunk
size in the `WriteIter`.  It’s set to the maximum size of a chunk that
can be sent in a transaction if no other instructions are sent.

Secondly, introduce `chunk_size` method to allow safely changing the
chunk size.

Thirdly, change `WriteIter::new` so that it length-prefixes the data
before writing it to the account.  For cases where user doesn’t want
the prefix to be included, add `WriteIter::new_raw` constructor.

Lastly, fix seed length checking in the `WriteIter` constructor.  Due
to optimisation of the write-account program, the seed can be at most
31 bytes but the constructor allowed 32-byte long seeds.